headers/deps: net/ip6: Factor out <net/ip6_route_api_mtu.h> from <net/ip6_route.h>
Decouple <net/ip6_route.h> from <linux/netdevice_api.h> by moving
the method that relies on it into its own header, and including
that header in the 3 .c files that use it.

In a distro build this reduces the number of <linux/netdevice_api.h>
inclusions by over 650 instances.

Signed-off-by: Ingo Molnar <mingo@kernel.org>