headers/deps: locking/spinlocks: Create a standard <linux/spinlock[_types|_api].h> header split
Currently <linux/spinlock.h> is the API header, and <linux/spinlock_types.h>
is the types header:

          --------------------------------------------------------------------
          | Combined, preprocessed C code size of header, without line markers,
          | with comments stripped:
          -------------------------
          | #include <linux/spinlock.h>                    | LOC:  8376 | headers: 215 |
          | #include <linux/spinlock_types.h>              | LOC:   455 | headers:  42 |

Move <linux/spinlock.h> to <linux/spinlock_api.h>, and create a default
<linux/spinlock.h> that includes both headers for the time being.

This is similar to how other _types/_api header splits are structured,
and is going to allow the broader use of <linux/spinlock_types.h>,
reducing header footprint in the kernel.

For this commit, there's no change in functionality or header footprint.

Signed-off-by: Ingo Molnar <mingo@kernel.org>