[BHF] ci: Add linux-image-bhf metapackages
Upstream uses a version scheme based on tags. An official release is
created on a tagged commit. This approach is not really feasible for us.
We have our CI in place and we want to test the changes before we decide
if it's a new release or not. This decision shouldn't be based on any
tags. For that reason, we're using the pipeline id for version
identification as it's guaranteed to monotonically increase.

User mostly shouldn't use a specific kernel version. They should update
their kernel if a new one is available. For an automated update when
running apt upgrade, we need a stable package name. Therefore, we're
using a meta package pointing to the latest kernel.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>