ci: use bhf version scheme
Upstream uses a version scheme based on tags. An official release is
created on a tagged commit. This approach is not really feasible for us.
We have our CI in place and we want to test the changes before we decide
if it's a new release or not. This decision shouldn't be based on any
tags. For that reason, we're using the pipeline id for version
identification as it's guaranteed to monotonically increase.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>