ci: build a real bhf meta package
Previously, our bhf meta package contained all binaries too. We didn't
notice it in CI because we're deleting the meta package and just install
the real package. However, you'll notice it, when trying to install the
meta package because apt will complain about conflicting files between
the files in the fake meta package and the depending package.

Fixes: 41b51e69c0ba4807d1ec2cc54f7a9ef8ac76d5f3
Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>