[BHF]: add arm64 kernel config
Kernel config is currently based on:
 - upstream debian config
 - make oldconfig
 - mergeconfig with defconfig for xilinx zynqmp

arm64: enable EFI_ZBOOT

u-boot and systemd boot don't support extracting compressed linux kernels.
This was recently introduced as a feature of the linux EFI stub and is
now available for all architectures.

arm64: use zstd compression for kernel

zstd is usually a better tradeoff than gzip so let's use it for arm64
where we have both slow storage and slow compute (on cx8200). This
tradeoff might be different on x86 and on different devices, so we
should actually benchmark these compression methods and levels if we
want to optimize for faster boot or smaller size at some point.

Reviewed-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>