[BHF]: add arm64 kernel config
This config was created from scratch.

We want to use preempt rt and support zynqmp devices.

Generated using:
run docker on cx8200
save output of lsmod to file (while docker is running)

config from debian linux 6.5-rt
make olddefconfig
make LSMOD=lsmod-8200 localmodconfig
make menuconfig
 - lima drm
 - zymqmp phy
 - zynqpm dpdma
 - zynqmp dp
 - tun
 - disable CONFIG_DEBUG_INFO_BTF (needs additional packages to build,
   maybe revisit)
 - enable HID

Disable psci cpuidle:
 We don't need it and it could cause latencies

Disable cpufreq:
 There's a performance regression with cpufreq in linux 6.8. We also
 don't want to scale cpu frequencies anyways, so disable it completely.
 https://patchwork.kernel.org/project/linux-acpi/patch/20231211104855.558096-4-vincent.guittot@linaro.org/
 https://lore.kernel.org/lkml/20240117190545.596057-1-vincent.guittot@linaro.org/

Enable support for multitouch panels:
 Added the configuration and tested multitouch  in chromium on sway
 on Beckhoff CP3918.
 Not sure if this is reasonable but it seems to work

Enable INPUT_EVDEV:
 We need this for input support in modern userspace

Enable iio-hwmon and xilinx-ams for zynqmp:
 We can read temperatures with this on CX9240 and CX8200

Enable display connector:
 We need it for zynqmp-dpsub

Set cma size to 128M:
 We need it for graphics on zynqmp.
 see: https://github.com/NixOS/nixpkgs/issues/260222

New since v6.9:
===============
[BHF] config: make olddefconfig for 6.9
We got a compile error in one of the renesas drivers. The
easiest fix was to disable support for that SOC.

[BHF] config: arm64: disable CONFIG_RANDOMIZE_KSTACK_OFFSET
It can currently cause random latencies on system calls, which is not
the kind of randomness we want. This seems to be a specific problem on
arm64.
Link: https://lore.kernel.org/linux-kernel/CAHmME9pYS=YZM1b0ftrjj_qjz_U4V8PsiLzB15JmBRr39Av2Ng@mail.gmail.com/T/

[BHF] arm64: enable MTD and spi nor flash support
We're using these to store factory and device data like mac addrs.
CONFIG_MTD_SPI_NOR=m for issi,is25lp016d on cx8200 and cx9240
As all of this is connected to qspi we also enable qspi support.

[BHF] config: arm64: enable latency tracers
At least SteffenD used these with ftrace to look for latencies.

[BHF] Enable netfilter modules for arm64
At the moment we have no firewall on arm64 and nftables fail to load
with:
  /etc/nftables.conf:9:5-12: Error: Could not process rule: No such file or directory
     ct state {established, related} accept
     ^^^^^^^^

We think this was caused by missing netfilter modules in the kernel so
we copy & pasted the netfilter part from config-amd64 into config-arm64.

[bhf] config: arm64: change sdhci driver to be built in
This is neccessary to boot without initramfs on zynqmp devices

[bhf] config: arm64: disable raid benchmarks
These are fairly slow on our devices and we're unlikely to use raid with
only one storage slot.

[bhf] config: arm64: switch from built-in ext4 to built-in btrfs
We use root on btrfs by default and want to be able to boot without
initramfs sometimes.

[bhf] arm64: enable gpio leds and triggers
We have some gpio controlled leds on our zynqmp devices.

[bhf] arm64: add scsi + usb storage support
scsi is needed for usb-attached-scsi, usb storage is needed for usb
sticks.

[BHF] arm64: enable onboard hub
Necessary for usb reset on boot on cx9240

[BHF] arm64: enable CONFIG_IP_PNP_DHCP
This enables automatic configuration of IP addresses of devices and
of the routing table during kernel boot, based on either information
supplied on the kernel command line or by BOOTP or RARP protocols.
You need to say Y only for diskless machines requiring network
access to boot (in which case you want to say Y to "Root file system
on NFS" as well), because all other machines configure the network
in their startup scripts.

[BHF] arm64: buildin BROADCOM_PHY/MACB
This will be required to enable ethernet communcation when booting the
kernel over nfs on zynqmp devices (i.e. CX82xx, CX9240).

[BHF] arm64: builtin ZYNQMP_GQSPI/MTD_SPI_NOR
The NOR flash stores the devices MAC addresses. U-Boot might write the
MAC addresses into the device tree already, so using something like:

	fw_devlink=permissive

should also work. However the cleaner solution might be to enable the
kernel to read the flash itself and avoid to ignore such errors.

[BHF] arm64: enable CONFIG_NFS_V3/CONFIG_ROOT_NFS
This allows to network-boot over NFS.

[BHF] config: enable XFS module
We have at least one customer, who wants to mount XFS partitions. So
here we go and add the module to our amd64 and arm64 kernel.

Signed-off-by: Daniel Kerkhoff <d.kerkhoff@beckhoff.com>
Signed-off-by: Julian Heuking <J.Heuking@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>
Signed-off-by: Steffen Dirkwinkel <s.dirkwinkel@beckhoff.com>

New since v6.11:
===============

[BHF] config-amd64: build ppp as module

We don't need it for booting so it's fine to have this as a module.
arm64 oldconfig

[BHF] config-arm64: build in MTD_OF_PARTS

since we build in spi nor and mtd support we had mtd loaded without
parsing of mtd partitions. Because of this we were missing nvmem-cells
files in sysfs (/sys/bus/nvmem/devices/mtd0/cells/...)

(cherry picked from commit cfb5bb451c0e0e118e569a7aa64f1c1c2b44210f)

[BHF] config-arm64: copy ppp options from amd64

This is used by some vpn software.

Fixes: https://azdevops01.beckhoff.com/GSW01/TcOs/_workitems/edit/209082

[BHF] config-arm64: disable CONFIG_IRQSOFF_TRACER

If the parameter CONFIG_IRQSOFF_TRACER is disabling the min time of the cyclictest is around 5000us lower.
The reason is that by enable CONFIG_IRQSOFF_TRACER parameter, a function are compiled in the trace module of the kernel which are trace the global enabling and disabling of the irq with separated trace message.

[BHF] config-arm64: enable virtio-blk and virtio-net driver

We want to run TC/LUR tests in a Qemu VM. Our Qemu VM will pass disks as
virtio-blk devices to the VM. Additionally, the network devices are
using virtio-net.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>

[BHF] config-arm64: Add gpio charger for 1s ups
Add the module driver gpio-charger for the support of the 1s ups from the cx9240 and cx82xx in TwinCAT

[BHF] config-arm64: make oldconfig

[BHF] config-arm64: add usb serial drivers to match amd64

USB support on arm64 and amd64 should match. Perhaps we should take a
look at device support in distro kernels and import it from someone.

note:
CONFIG_USB_SERIAL_WHITEHEAT selects CONFIG_USB_EZUSB_FX2=m

Reported-by: Daniel Kerkhoff <d.kerkhoff@beckhoff.com>
Reported-by: Jan-Alrik Zipter <J.Zipter@beckhoff.com>

[BHF] enable CONFIG_USB_ONBOARD_DEV

CONFIG_USB_ONBOARD_HUB was renamed in commit 31e7f6c015d9eb35e77ae9868801c53ab0ff19ac
We didn't catch that and this broke usb support on cx9240.

I'll also add a test for this now.

New since v6.12:
===============

Make olddefconfig for 6.12 arm64
Enable crypto required by iwd arm64
Enable crypto for EAP-PEAP-MSCHAPV2 arm64
Enable DM_CRYPT arm64