arm64: use zstd compression for kernel
zstd is usually a better tradeoff than gzip so let's use it for arm64
where we have both slow storage and slow compute (on cx8200). This
tradeoff might be different on x86 and on different devices, so we
should actually benchmark these compression methods and levels if we
want to optimize for faster boot or smaller size at some point.