[BHF] debian: add initramfs-tools as dependency
mmdebstrap supports hooks to modify the rootfs while building it. This
allows us to add our Beckhoff package repository to mmdebstrap and
installing our Beckhoff packages with mmdebstrap directly.
Unfortunately, the update-initramfs trigger is broken when the kernel is
installed before initramfs-tools. Therefore, we have to force apt to
install the kernel after initramfs-tools by adding a dependency to our
kernel.

We've tested if Recommends is sufficient because it would allow user to
uninstall initramfs-tools without uninstalling the kernel. It's not
sufficient. However, at the moment we always need an initramfs to boot
our ARM devices. If that changes in future, we can easily remove the
dependency.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>