[BHF] arm64: enable CONFIG_IP_PNP_DHCP
This enables automatic configuration of IP addresses of devices and
of the routing table during kernel boot, based on either information
supplied on the kernel command line or by BOOTP or RARP protocols.
You need to say Y only for diskless machines requiring network
access to boot (in which case you want to say Y to "Root file system
on NFS" as well), because all other machines configure the network
in their startup scripts.

Signed-off-by: Julian Heuking <J.Heuking@beckhoff.com>