[BHF] debian: add initramfs-tools as dependency for all architectures
We previously only added this for arm64. Our memstick installer depends
on it for the read only init (it expects /sys to be mounted, which the
initramfs seems to mount for us). We add it for all architectures now.
This somehow worked for amd64 by chance before, as initramfs-tools was
always set up before the kernel.

Investigated-by: Julian Heuking <j.heuking@beckhoff.com>