[BHF] config-arm64: disable CONFIG_IRQSOFF_TRACER
If the parameter CONFIG_IRQSOFF_TRACER is disabling the min time of the cyclictest is around 5000us lower.
The reason is that by enable CONFIG_IRQSOFF_TRACER parameter, a function are compiled in the trace module of the kernel which are trace the global enabling and disabling of the irq with separated trace message.