[BHF]: ixgbe: Ignored Mod_ABS interrupt signal for CX20X2 as workaround
In the network driver ixgbe from intel ignore the
Mod_ABS interrupt signal (Module absent, connected)
for the Beckhoff Automation CX20X2.
The reason is, that hardware of the CX20x2 has a emv bug, so that
interrupt signal are set by the crosstalk other clk lines (power supply).
The devices are currently unusable for us with this interrupt
(network resets constantly), but work fine if we ignore it,
so we ignore it until our HW department has a better fix.