Revert "ci: build with our llvm toolchain"
We've recently found that this breaks out of tree modules
as these would try to use llvm only options (‘-mretpoline-external-thunk’)
during builds.

This was already noticed by others and things like dkms switch
automatically:
https://github.com/dell/dkms/pull/169/files

But right now we don't yet ship the llvm package to customers and this
also breaks our ccat pipeline, so we revert the change until we have a
fix in the ccat test.

This reverts commit caa47dd3fa634ebf881bc6e6fcc6cb38cc2f2181.