[BHF]: debian: Allow either dracut or initramfs-tools
To adopt more of the systemd ecosystem, we make the switch from a
busybox-based initrd with initramfs-tools to a systemd-based initrd with
dracut.
To make migration easier and avoid breaking existing application that
rely on initramfs-tools, we allow for either one to be installed.
Apt will continue to install initramfs-tools by default unless dracut is
explicitly installed or pulled in via another dependency.
This helps us later when making changes to debian-install: instead of
checking whether initramfs-tools or dracut is installed, we can install
dracut explicitly and always be sure that it's the only initrd
installed.