[BHF]: debian: build unsigned unified kernel image
As part of implementing secure and measured boot, we take the first step
towards migrating to Unified Kernel Images by publishing unsigned UKIs as a
debian package.
Similar to our regular kernel packages, those will be available both under
their explicit version (linux-uki-$kver-unsigned) and as a meta-package
pointing to the latest one (linux-uki-bhf-unsigned).
It is important to note that UKIs make it harder to edit the kernel cmdline.
Any changes should be made by building (and later signing) a UKI addon
containing the extra cmdline parameters.
For now we include addons for TcCoreConf configs with up to 64 shared cores.