[BHF]: Add bbapi-sups device
Add the sups device from the BBAPI[1] project as "bbapi-sups".
The bbapi-sups implements a GPIO that exposes the sups status
to user space.
Expand "bbapi-sups" so that the gpio-charger device is created using
the GPIO provided by sups. This ensures that in user space the SUPS
status interface (sys/class/power_supply/ups/status) is the same on
both Beckhoff ARM and x86_64 Linux CX devices.

[1] https://git.beckhoff.dev/beckhoff/BBAPI