config-amd64: load i915 as module
To avoid including the intel gpu firmware in the initrd and thus making the UKI
even larger, we defer it until udev loads it later, after the root filesystem
is ready.
We already do it this way for the amdgpu firmware, so we only have to configure
it for intel's i915 drivers.