[BHF]: config-amd64: set default ASPM policy to performance
We have some older bios revisions were this isn't the default in bios
and may cause higher latencies. We don't really ever want to use aspm,
so we can set the default policy to performance in the kernel.
This doesn't disable aspm in all cases, as sometimes it may be enabled
by the bios without allowing changes to the kernel.

Fixes: https://azdevops01.beckhoff.com/GSW01/TcOs/_workitems/edit/191955/