[BHF]: fix cross-compiling dependency for binutils
binutils arm64 can't be installed alongside binutils amd64 and this
breaks cross building. If anything binutils-multiarch may be correct.
We fix it for now by just using the native binutils version always.

When cross-compiling and building the Debian package of the Linux kernel
for an arm64 system, the native ("host") version of the binutils package
must be installed.
If the arm64 version of the binutils package is installed on the
cross-compiling system, other important packages required for the kernel
build —such as gcc-14-aarch64-linux-gnu, gcc-aarch64-linux-gnu, and
others— are removed during installation.