debian: restore isolated cores on uki upgrade
With the upgrade from regular kernels to unified kernel images, the
mechanism to configure isolated cores had to change as well [1].
This lead to the user losing their core configuration after migrating.
To fix this, we make the postinstall script apply the "legacy" isolated
core configuration to the installed UKI.
For that, we can look at one of two files:

/etc/kernel/cmdline:

/etc/kernel/cmdline gets written immediately after calling TcCoreConf,
without the need to reboot, but introduces the risk of "resurrecting" an
older core configuration that has been written (long) before the
migration to UKIs took place.
Normally, this shouldn't happen, because even removing the shared core
configuration would cause a UKI-addon to get copied to the ESP (with the
shared cores equal to the effective cores) and the migration to get
skipped.
Only when the user explicitly removes their UKI-addon from the ESP,
would this cause the aforementioned "resurrection" of an outdated
/etc/kernel/cmdline.

/proc/cmdline:

/proc/cmdline shows the currently active kernel parameters and therefore
only updates on reboots. This gets around the "resurrection"-problem,
but introduces a new edge case, which can be reproduced like this:

1. Run TcCoreConf --set-shared=X
2. apt upgrade from a regular kernel to UKIs
3. Reboot the system

This chain of events would cause the configuration from step 1 to be
ignored entirely, because the newly booted UKI doesn't read the updated
/etc/kernel/cmdline. The migration script would look at an outdated
cmdline at the time of the upgrade.

For the time being we pick /etc/kernel/cmdline, as the user should never
remove UKI addons manually from the ESP. The "ressurection"-edge-case is
therefore a lot less likely.

[1]: https://azdevops01.beckhoff.com/GSW01/TwinCAT%20Base/_git/TcBase/pullrequest/106494