[BHF]: publish kernel with our package
In commit 29c78c4525e6f982e0f690009e1bd3ab6451cd7a [1] we introduced
UKIs as our new boot enviroment for Linux. During this change we broke
guestfish because it has a dependency to the upstream linux-image-amd64.

The supermin which builds the enviroment for guestfish does not support
UKIs yet. In the future we should add the support for UKIs to supermin.
This however will take time and I want to publish luemuctl in the near
future.
Thus I want to publish the raw kernel with our package aswell.
supermin is searching in/lib/modules/*/vmlinuz for a suitable kernel.