[BHF]: add CONFIG_SCSI_VIRTIO as kernel module
In commit 29c78c4525e6f982e0f690009e1bd3ab6451cd7a [1] we introduced
UKIs as our new boot enviroment for Linux. During this change we broke
guestfish because it has a dependency to the upstream linux-image-amd64.

Right now our kernel image (vmlinuz) is not compatible with
libguestfs-tools.
In the next commits we want to make our kernel package compatible with
libguestfs-tools.

To access the scsi disk, libguestfs-tools requires CONFIG_SCSI_VIRTIO to
be available. We add it as a module because we don't want to bloat our
kernel image with unnecessary drivers.

[1] https://git.beckhoff.dev/beckhoff/linux/-/commit/29c78c4525e6f982e0f690009e1bd3ab6451cd7a